Processing-Embedding-Service/requirements.txt
# --- Core ---
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# --- AWS ---
boto3>=1.36.0
aioboto3>=13.3.0

# --- Text extraction ---
pdfplumber>=0.11.0
python-docx>=1.1.0
openai-whisper>=20231117

# --- Embeddings ---
openai>=1.60.0
sentence-transformers>=3.3.0

# --- Vector stores ---
faiss-cpu>=1.9.0
pinecone-client>=5.0.0

# --- Utilities ---
tiktoken>=0.8.0
aiosqlite>=0.20.0

# --- Testing ---
pytest>=8.3.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
moto[s3,sqs]>=5.0.0
